Moved free-trial.ts to billing
