+++ apps/web/src/app/vault/individual-vault/vault-banners/vault-banners.component.ts
@@ -9,7 +9,7 @@ import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.servic
 import { BannerModule } from "@bitwarden/components";
 
 import { VerifyEmailComponent } from "../../../auth/settings/verify-email.component";
-import { FreeTrial } from "../../../core/types/free-trial";
+import { FreeTrial } from "../../../billing/types/free-trial";
 import { SharedModule } from "../../../shared";
 
 import { VaultBannersService, VisibleVaultBanner } from "./services/vault-banners.service";
