+++ apps/web/src/app/vault/individual-vault/vault.component.ts
@@ -85,7 +85,7 @@ import {
 } from "@bitwarden/vault";
 
 import { TrialFlowService } from "../../billing/services/trial-flow.service";
-import { FreeTrial } from "../../core/types/free-trial";
+import { FreeTrial } from "../../billing/types/free-trial";
 import { SharedModule } from "../../shared/shared.module";
 import { AssignCollectionsWebComponent } from "../components/assign-collections";
 import {
